requirements.txt
alembic==1.13.1
annotated-types==0.7.0
dnspython==2.6.1
email_validator==2.1.1
grpcio==1.64.1
grpcio-tools==1.64.1
idna==3.7
Mako==1.3.5
MarkupSafe==2.1.5
passlib==1.7.4
protobuf==5.27.1
psycopg2-binary==2.9.9
pydantic==2.7.3
pydantic_core==2.18.4
PyJWT==2.8.0
SQLAlchemy==2.0.30
typing_extensions==4.12.1
